[storage] Refactor file catalog test
